extract to file, add sources
